feat: coarse moving theme related files
