+++ packages/mux-player/src/media-chrome/time-display.ts
